requirements.txt
fastapi
uvicorn
websockets
httpx
python-dotenv
jinja2

